V 2.0
Code is refactored structure changed
