+++ Constants/Options.py
@@ -1,0 +1,4 @@
+teams= "ms_teams_exports_usernev_dll.txt"
+onedrive="onedrive_exports_version_dll.txt"
+
+file_options = {"teams": teams, "onedrive": onedrive}
