+++ Constants/Paths.py
@@ -1,0 +1,1 @@
+folder_path="./Dll_Exports/"
