+++ Templates/Banner.py
@@ -1,0 +1,9 @@
+print(r"""
+  ______  __    __   __  .___  ___.  _______ .______          ___      
+ /      ||  |  |  | |  | |   \/   | |   ____||   _  \        /   \     
+|  ,----'|  |__|  | |  | |  \  /  | |  |__   |  |_)  |      /  ^  \    
+|  |     |   __   | |  | |  |\/|  | |   __|  |      /      /  /_\  \   
+|  `----.|  |  |  | |  | |  |  |  | |  |____ |  |\  \----./  _____  \  
+ \______||__|  |__| |__| |__|  |__| |_______|| _| `._____/__/     \__\ 
+                                                                       
+""")
